+++ src/AddressContainer.test.js
@@ -4,10 +4,16 @@ import AddressContainer from './AddressContainer'
 import address from './__mocks__/newAddress'
 import usePostalCode from './country/__mocks__/usePostalCode'
 import postalCodeAutoCompleteAddress from './postalCodeAutoCompleteAddress'
-
-/* #FIXME when enzyme adds suport for context */
+import PostalCodeGetter from './PostalCodeGetter'
 
 jest.mock('./postalCodeAutoCompleteAddress')
+
+const descendToChild = wrapper =>
+  wrapper
+    .children()
+    .children()
+    .children()
+    .children()
 
 describe('AddressContainer', () => {
   const accountName = 'qamarketplace'
@@ -17,7 +23,8 @@ describe('AddressContainer', () => {
     country: { value: 'BRA' },
   }
 
-  xit('should render its children', () => {
+  it('should render its children', () => {
+    // Arrange
     const wrapper = mount(
       <AddressContainer
         cors
@@ -29,16 +36,18 @@ describe('AddressContainer', () => {
         <span className="unique" />
       </AddressContainer>,
     )
-    expect(wrapper.find('.unique')).toHaveLength(1)
+
+    // Act
+    const result = wrapper.find('.unique')
+
+    // Assert
+    expect(result).toHaveLength(1)
   })
 
-  xit('should handle the validation before calling onChangeAddress', () => {
+  it('should handle the validation before calling onChangeAddress', () => {
+    // Arrange
     const handleAddressChange = jest.fn()
-    const children = onChangeAddress => {
-      onChangeAddress({ city: { value: 'Rio de Janeiro' } })
-      return <span />
-    }
-    mount(
+    const wrapper = mount(
       <AddressContainer
         cors
         accountName={accountName}
@@ -46,22 +55,25 @@ describe('AddressContainer', () => {
         onChangeAddress={handleAddressChange}
         rules={usePostalCode}
       >
-        {children}
+        <PostalCodeGetter rules={usePostalCode} />
       </AddressContainer>,
     )
+
+    // Act
+    const onChangeAddress = descendToChild(wrapper).props().onChangeAddress
+    onChangeAddress({ city: { value: 'Rio de Janeiro' } })
+
+    // Assert
     expect(handleAddressChange).toHaveBeenCalledWith({
       ...addressWithCountry,
       city: { value: 'Rio de Janeiro', valid: true },
     })
   })
 
-  xit('should short-circuit when country changes', () => {
+  it('should short-circuit when country changes', () => {
+    // Arrange
     const handleAddressChange = jest.fn()
-    const children = jest.fn(onChangeAddress => {
-      onChangeAddress({ country: { value: 'ECU' } })
-      return <span />
-    })
-    mount(
+    const wrapper = mount(
       <AddressContainer
         cors
         accountName={accountName}
@@ -69,9 +81,15 @@ describe('AddressContainer', () => {
         onChangeAddress={handleAddressChange}
         rules={usePostalCode}
       >
-        {children}
+        <PostalCodeGetter rules={usePostalCode} />
       </AddressContainer>,
     )
+
+    // Act
+    const onChangeAddress = descendToChild(wrapper).props().onChangeAddress
+    onChangeAddress({ country: { value: 'ECU' } })
+
+    // Assert
     expect(handleAddressChange).toHaveBeenCalledWith({
       ...addressWithCountry,
       country: { value: 'ECU' },
@@ -83,13 +101,10 @@ describe('AddressContainer', () => {
       postalCodeAutoCompleteAddress.mockClear()
     })
 
-    xit('should auto complete postal code when postal code is valid', () => {
+    it('should auto complete postal code when postal code is valid', () => {
+      // Arrange
       const handleAddressChange = jest.fn()
-      const children = jest.fn(onChangeAddress => {
-        onChangeAddress({ postalCode: { value: '22231000' } })
-        return <span />
-      })
-      mount(
+      const wrapper = mount(
         <AddressContainer
           cors
           accountName={accountName}
@@ -97,19 +112,22 @@ describe('AddressContainer', () => {
           onChangeAddress={handleAddressChange}
           rules={usePostalCode}
         >
-          {children}
+          <PostalCodeGetter rules={usePostalCode} />
         </AddressContainer>,
       )
+
+      // Act
+      const onChangeAddress = descendToChild(wrapper).props().onChangeAddress
+      onChangeAddress({ postalCode: { value: '22231000' } })
+
+      // Assert
       expect(postalCodeAutoCompleteAddress).toHaveBeenCalled()
     })
 
-    xit('should not auto complete postal code when postal code is invalid', () => {
+    it('should not auto complete postal code when postal code is invalid', () => {
+      // Arrange
       const handleAddressChange = jest.fn()
-      const children = jest.fn(onChangeAddress => {
-        onChangeAddress({ postalCode: { value: '222' } })
-        return <span />
-      })
-      mount(
+      const wrapper = mount(
         <AddressContainer
           cors
           accountName={accountName}
@@ -117,21 +135,22 @@ describe('AddressContainer', () => {
           onChangeAddress={handleAddressChange}
           rules={usePostalCode}
         >
-          {children}
+          <PostalCodeGetter rules={usePostalCode} />
         </AddressContainer>,
       )
+
+      // Act
+      const onChangeAddress = descendToChild(wrapper).props().onChangeAddress
+      onChangeAddress({ postalCode: { value: '222' } })
+
+      // Assert
       expect(postalCodeAutoCompleteAddress).not.toHaveBeenCalled()
     })
 
-    xit('should not auto complete postal code when postal code was auto completed by geolocation', () => {
+    it('should not auto complete postal code when postal code was auto completed by geolocation', () => {
+      // Arrange
       const handleAddressChange = jest.fn()
-      const children = jest.fn(onChangeAddress => {
-        onChangeAddress({
-          postalCode: { value: '22231000', geolocationAutoCompleted: true },
-        })
-        return <span />
-      })
-      mount(
+      const wrapper = mount(
         <AddressContainer
           cors
           accountName={accountName}
@@ -139,21 +158,24 @@ describe('AddressContainer', () => {
           onChangeAddress={handleAddressChange}
           rules={usePostalCode}
         >
-          {children}
+          <PostalCodeGetter rules={usePostalCode} />
         </AddressContainer>,
       )
+
+      // Act
+      const onChangeAddress = descendToChild(wrapper).props().onChangeAddress
+      onChangeAddress({
+        postalCode: { value: '22231000', geolocationAutoCompleted: true },
+      })
+
+      // Assert
       expect(postalCodeAutoCompleteAddress).not.toHaveBeenCalled()
     })
 
-    xit('should not auto complete postal code when prop autoCompletePostalCode is false', () => {
+    it('should not auto complete postal code when prop autoCompletePostalCode is false', () => {
+      // Arrange
       const handleAddressChange = jest.fn()
-      const children = jest.fn(onChangeAddress => {
-        onChangeAddress({
-          postalCode: { value: '22231000' },
-        })
-        return <span />
-      })
-      mount(
+      const wrapper = mount(
         <AddressContainer
           cors
           accountName={accountName}
@@ -162,9 +184,15 @@ describe('AddressContainer', () => {
           autoCompletePostalCode={false}
           rules={usePostalCode}
         >
-          {children}
+          <PostalCodeGetter rules={usePostalCode} />
         </AddressContainer>,
       )
+
+      // Act
+      const onChangeAddress = descendToChild(wrapper).props().onChangeAddress
+      onChangeAddress({ postalCode: { value: '22231000' } })
+
+      // Assert
       expect(postalCodeAutoCompleteAddress).not.toHaveBeenCalled()
     })
   })