+++ src/geolocation/index.js
@@ -1,3 +1,3 @@
-export GoogleMapsContainer from './GoogleMapsContainer'
-export GeolocationInput from './GeolocationInput'
-export Map from './Map'
+export { default as GoogleMapsContainer } from './GoogleMapsContainer'
+export { default as GeolocationInput } from './GeolocationInput'
+export { default as Map } from './Map'