+++ src/index.js
@@ -1,17 +1,18 @@
 // Components
-export CountrySelector from './CountrySelector'
-export AddressForm from './AddressForm'
-export AddressSummary from './AddressSummary'
-export PostalCodeGetter from './PostalCodeGetter'
-export AddressContainer from './AddressContainer'
-export AutoCompletedFields from './AutoCompletedFields'
-export AddressRules from './AddressRules'
-export AddressSubmitter from './AddressSubmitter'
+export { default as CountrySelector } from './CountrySelector'
+export { default as AddressForm } from './AddressForm'
+export { default as AddressSummary } from './AddressSummary'
+export { default as PostalCodeGetter } from './PostalCodeGetter'
+export { default as AddressContainer } from './AddressContainer'
+export { default as AutoCompletedFields } from './AutoCompletedFields'
+export { default as AddressRules } from './AddressRules'
+export { default as AddressSubmitter } from './AddressSubmitter'
 
 // Constants
-export constants from './constants'
+export { default as constants } from './constants'
 
 // Helper Functions
 export { addValidation, removeValidation } from './transforms/address'
 export { isValidAddress, validateField } from './validateAddress'
 export { injectRules } from './addressRulesContext'
+export { injectAddressContext } from './addressContainerContext'